Add fsm to balancing subsystem

+++ mainboard/Inc/peripherals/ltc6813_utils.h
@@ -59,8 +59,8 @@ void ltc6813_check_voltage(uint16_t volts, uint8_t index);
  */
 void ltc6813_check_temperature(uint16_t temps, uint8_t index);
 
-void ltc6813_set_dcc(uint8_t indexes[], uint8_t cfgar[8], uint8_t cfgbr[8]);
-void ltc6813_set_balancing(SPI_HandleTypeDef *hspi, uint8_t *indexes, int dcto);
+void ltc6813_set_dcc(uint16_t indexes[], uint8_t cfgar[8], uint8_t cfgbr[8]);
+void ltc6813_set_balancing(SPI_HandleTypeDef *hspi, uint16_t *indexes, int dcto);
 
 /**
  * @brief	This function is used to convert the 2 byte raw data from the LTC68xx to a 16 bit unsigned integer
